pyproject.toml
[project]
name = "pesuacademy"
version = "1.0.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
license = { text = "MIT" }
authors = [
    { name = "Aditeya Baral", email = "aditeya.baral@gmail.com" },
    { name = "Samarth Mohan", email = "samarth.lina@gmail.com" }
]
dependencies = [
    "beautifulsoup4>=4.13.4",
    "httpx>=0.28.1",
    "pydantic>=2.11.7",
    "python-dotenv>=1.1.1",
    "selectolax>=0.3.32",
]

[project.optional-dependencies]
pre-commit = [
    "pre-commit>=4.2.0",
    "pytest>=8.4.1",
    "pytest-asyncio>=1.1.0",
    "pytest-cov>=6.2.1",
    "ruff>=0.12.7",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["."]

[tool.pytest.ini_options]
pythonpath = ["."]
markers = [
    "secret_required: marks tests that require secrets or environment variables (e.g. TEST_PRN, TEST_PASSWORD)"
]

[tool.ruff.lint]
select = [
    "E", # Pycodestyle errors
    "F", # Pyflakes errors
    "W", # Pycodestyle warnings
    "UP", # Pyupgrade rules
    "I", # Import related rules
    "C90", # Complexity rules
    "D", # Documentation rules
    "ANN", # Annotations rules
    "TYP", # Type checking rules
    "N", # Naming rules
    "Q000", # Quotation style rules
    "RET", # Return type rules
]

[tool.ruff.lint.pydocstyle]
convention = "google"

[tool.ruff]
line-length = 120
target-version = "py312"
exclude = [
    "tests/",
    "scripts/",
    "examples/"
]

[tool.ruff.format]
quote-style = "double"
docstring-code-format = true
